% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{as_keyed_tibble1}
\alias{as_keyed_tibble1}
\title{Convert \code{x} to a \code{keyed_tibble1}}
\usage{
as_keyed_tibble1(x, key_colnames = 1, ...)
}
\arguments{
\item{key_colnames}{dummy attr, just here to test how it moves around}
}
\description{
Convert \code{x} to a \code{keyed_tibble1}
}
