% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{as_nodupe_tibble}
\alias{as_nodupe_tibble}
\title{Convert \code{x} to a \code{nodupe_tibble}}
\usage{
as_nodupe_tibble(x, ...)
}
\arguments{
\item{my_attr}{dummy attr, just here to test how it moves around}
}
\description{
Convert \code{x} to a \code{nodupe_tibble}
}
