% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{check_df_keyed_tibble1_compatible}
\alias{check_df_keyed_tibble1_compatible}
\title{Is / why isn't data.frame/subclass \code{x} compatible with keyed_tibble1 invariants}
\usage{
check_df_keyed_tibble1_compatible(x, key_colnames)
}
\arguments{
\item{x}{data.frame or subclass}
}
\value{
TRUE or str
}
\description{
Is / why isn't data.frame/subclass \code{x} compatible with keyed_tibble1 invariants
}
