% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{check_df_nodupe_tibble_compatible}
\alias{check_df_nodupe_tibble_compatible}
\title{Is / why isn't data.frame/subclass \code{x} compatible with nodupe_tibble invariants}
\usage{
check_df_nodupe_tibble_compatible(x, my_attr)
}
\arguments{
\item{x}{data.frame or subclass}
}
\value{
TRUE or str
}
\description{
Is / why isn't data.frame/subclass \code{x} compatible with nodupe_tibble invariants
}
