% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{decay_nodupe_tibble}
\alias{decay_nodupe_tibble}
\title{Remove nodupe_tibble class & attrs:}
\usage{
decay_nodupe_tibble(x, ...)
}
\description{
Remove nodupe_tibble class & attrs:
}
\keyword{internal}
