% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{ensure_decayed_keyed_tibble1}
\alias{ensure_decayed_keyed_tibble1}
\title{Remove keyed_tibble1 class & attrs if present:}
\usage{
ensure_decayed_keyed_tibble1(x, ...)
}
\description{
Remove keyed_tibble1 class & attrs if present:
}
\keyword{internal}
