% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{ensure_new_keyed_tibble1}
\alias{ensure_new_keyed_tibble1}
\title{Ensure keyed_tibble1 is head class & attrs as given}
\usage{
ensure_new_keyed_tibble1(x, key_colnames)
}
\description{
Ensure keyed_tibble1 is head class & attrs as given
}
\keyword{internal}
