% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{ensure_new_nodupe_tibble}
\alias{ensure_new_nodupe_tibble}
\title{Ensure nodupe_tibble is head class & attrs as given}
\usage{
ensure_new_nodupe_tibble(x, my_attr = 1)
}
\description{
Ensure nodupe_tibble is head class & attrs as given
}
\keyword{internal}
