% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{maybe_decay_nodupe_tibble}
\alias{maybe_decay_nodupe_tibble}
\title{Validate whether x obeys invariants of nodupe_tibble; if not, decay to non-nodupe_tibble:}
\usage{
maybe_decay_nodupe_tibble(x)
}
\description{
Validate whether x obeys invariants of nodupe_tibble; if not, decay to non-nodupe_tibble:
}
\keyword{internal}
