% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{maybe_ensure_decayed_keyed_tibble1}
\alias{maybe_ensure_decayed_keyed_tibble1}
\title{Validate whether x obeys invariants of keyed_tibble1; if not, decay to non-keyed_tibble1:}
\usage{
maybe_ensure_decayed_keyed_tibble1(x)
}
\description{
Validate whether x obeys invariants of keyed_tibble1; if not, decay to non-keyed_tibble1:
}
\keyword{internal}
