% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{maybe_new_keyed_tibble1}
\alias{maybe_new_keyed_tibble1}
\title{Ensure nodupe is head class & attrs set, if compatible with nodupe}
\usage{
maybe_new_keyed_tibble1(x, key_colnames)
}
\description{
Ensure nodupe is head class & attrs set, if compatible with nodupe
}
\keyword{internal}
