% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{maybe_new_keyed_tibble1_0}
\alias{maybe_new_keyed_tibble1_0}
\title{maybe_ensure_decayed_keyed_tibble1 if we know x is.data.frame}
\usage{
maybe_new_keyed_tibble1_0(x, key_colnames)
}
\description{
maybe_ensure_decayed_keyed_tibble1 if we know x is.data.frame
}
\keyword{internal}
