% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{maybe_new_nodupe_tibble}
\alias{maybe_new_nodupe_tibble}
\title{Ensure nodupe is head class & attrs set, if compatible with nodupe}
\usage{
maybe_new_nodupe_tibble(x, my_attr = 1)
}
\description{
Ensure nodupe is head class & attrs set, if compatible with nodupe
}
\keyword{internal}
