% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{maybe_new_nodupe_tibble0}
\alias{maybe_new_nodupe_tibble0}
\title{maybe_decay_nodupe_tibble if we know x is.data.frame}
\usage{
maybe_new_nodupe_tibble0(x, my_attr = 1)
}
\description{
maybe_decay_nodupe_tibble if we know x is.data.frame
}
\keyword{internal}
