% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodupe_tibble.R
\name{new_dedupe_tibble}
\alias{new_dedupe_tibble}
\title{Low-level constructor for nodupe_tibble; use as_nodupe_tibble or use carefully}
\usage{
new_dedupe_tibble(x, my_attr = 1)
}
\arguments{
\item{my_attr}{dummy attr, just here to test how it moves around}
}
\description{
This does not validate that there are no duplicate rows in \code{x}; use this only
if you have already verified that. For a constructor/converter that validates
that for you, use \code{as_nodupe_tibble} instead.
}
