% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed_tibble1.R
\name{new_keyed_tibble1}
\alias{new_keyed_tibble1}
\title{Low-level constructor for keyed_tibble1; use as_keyed_tibble1 or use carefully}
\usage{
new_keyed_tibble1(x, key_colnames)
}
\arguments{
\item{key_colnames}{chr}
}
\description{
This does not validate that there are no duplicate rows in \code{x}; use this only
if you have already verified that. For a constructor/converter that validates
that for you, use \code{as_keyed_tibble1} instead.
}
